Break down the app into components
