+++ src/components/Footer.jsx
@@ -1,0 +1,20 @@
+const Footer = () => {
+  return (
+    <footer className="text-center text-gray-500 text-[0.75rem]">
+      <p>
+        Songs I&apos;m listening to on{" "}
+        <a
+          href="https://open.spotify.com/user/utawon8gsgty2atyw86cw94ym"
+          target="_blank"
+          rel="noreferrer"
+          className="text-gray-400 hover:text-gray-300"
+        >
+          Spotify{" "}
+        </a>
+        :)
+      </p>
+    </footer>
+  );
+};
+
+export default Footer;
